project/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
flask-login==0.6.3
email-validator==2.1.0.post1
python-dotenv==1.0.0